refactor: 이벤트 리스너는 service 하위로 이동
